requirements.txt
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0




